fix #106: Jpa 관련 오류 수정
