+++ com.msa_delivery.delivery/src/main/java/com/msa_delivery/delivery/domain/repository/DeliveryManagerRepository.java
@@ -8,16 +8,15 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import java.util.List;
 import java.util.Optional;
 
 public interface DeliveryManagerRepository  extends JpaRepository<DeliveryManager, Long>, JpaDeliveryManagerRepository {
     Optional<DeliveryManager> findByIdAndIsDeleteFalse(Long deliveryManagerId);
 
-    Optional<DeliveryManager> findFirstByDeliveryManagerTypeAndOrderIdIsNullAndIsDeleteFalseOrderBySequenceAsc(DeliveryManagerType deliveryManagerType);
+    Optional<DeliveryManager> findFirstByTypeAndOrderIdIsNullAndIsDeleteFalseOrderBySequenceAsc(DeliveryManagerType type);
 
-    @Query("SELECT MAX(dm.sequence) FROM DeliveryManager dm WHERE dm.type = ? AND dm.isDelete = false")
-    Integer findMaxSequenceByType(DeliveryManagerType type);
+    @Query("SELECT MAX(dm.sequence) FROM DeliveryManager dm WHERE dm.type = :type AND dm.isDelete = false")
+    Integer findMaxSequenceByType(@Param("type") DeliveryManagerType type);
 
     @Modifying
     @Query("UPDATE DeliveryManager dm SET dm.sequence = dm.sequence + 1 WHERE dm.type = :type AND dm.sequence >= :sequence AND dm.isDelete = false")
