+++ com.msa_delivery.delivery/src/main/java/com/msa_delivery/delivery/infrastructure/repository/JpaDeliveryManagerRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.domain.Pageable;
 import java.util.UUID;
 
 public interface JpaDeliveryManagerRepository {
-    Page<DeliveryManagerDto> searchManagers(String serach, String type, UUID hubId, Integer sequenceMin, Integer sequenceMax,
+    Page<DeliveryManagerDto> searchManagers(String search, String type, UUID hubId, Integer sequenceMin, Integer sequenceMax,
                                             String createdFrom, String createdTo, Pageable pageable);
 }
